Revert "Release v0.5.0: Print writers, add capability to upsert cost explorer results to a Pinecone index. Involves a preprocessing step which embeds vector results. Change also includes removal of gpt summarization functionality (#58)"
This reverts commit 72d3159cfbcba1e37b2a90a544c097a2932c67a2.
Reverting commit
