+++ internal/core/domain/model/command_types.go
@@ -36,8 +36,6 @@ type CommandLineInput struct {
 	Metrics             []string
 	SortByDate          bool
 	OpenAIAPIKey        string
-	PineconeIndex       string
-	PineconeAPIKey      string
 }
 
 type FilterBySelections struct {
@@ -54,8 +52,6 @@ type PrintOptions struct {
 	OpenAIKey        string
 	Granularity      string
 	Metric           string
-	PineconeIndex    string
-	PineconeAPIKey   string
 }
 
 type ForecastCommandLineInput struct {
@@ -137,8 +133,6 @@ type CostAndUsageRequestType struct {
 	Metrics                    []string
 	SortByDate                 bool
 	OpenAIAPIKey               string
-	PineconeIndex              string
-	PineconeAPIKey             string
 }
 
 type CostAndUsageRequestWithResourcesType struct {
