+++ internal/core/handlers/aws/validator.go
@@ -1,4 +1,4 @@
-package commandline
+package aws
 
 import (
 	"github.com/cduggn/ccexplorer/internal/core/domain/model"
@@ -26,22 +26,20 @@ func ValidateInput(input model.CommandLineInput) error {
 	if !isValidPrintFormat {
 		return ValidationError{
 			Message: "Invalid print format. " +
-				"Please use one of the following: stdout, csv, chart, pinecone",
+				"Please use one of the following: stdout, csv, chart, gpt",
 		}
 	}
 
-	if input.PrintFormat == "pinecone" && input.OpenAIAPIKey == "" {
+	if input.PrintFormat == "gpt" && input.OpenAIAPIKey == "" {
 		return ValidationError{
-			Message: "OpenAI API key not set. " +
-				"Please set the OPENAI_API_KEY in the config file or environment variable",
+			Message: "OpenAI API key not set. Please set the open_ai_api_key in the config file or environment variable",
 		}
 	}
 
-	if input.PrintFormat == "pinecone" && input.OpenAIAPIKey != "" {
+	if input.PrintFormat == "gpt" && input.OpenAIAPIKey != "" {
 		if HasAccountInformation(input.GroupByDimension) {
 			return ValidationError{
-				Message: "Cannot use Pinecone with account information. " +
-					"Please remove the account dimension",
+				Message: "Cannot use GPT with account information. Please remove the account dimension",
 			}
 		}
 	}
@@ -101,7 +99,7 @@ func ValidateEndDate(endDate, startDate string) error {
 }
 
 func IsValidPrintFormat(f string) bool {
-	return f == "stdout" || f == "csv" || f == "chart" || f == "pinecone"
+	return f == "stdout" || f == "csv" || f == "chart" || f == "gpt"
 }
 
 func IsValidGranularity(g string) bool {
