+++ internal/core/presentation/interfaces.go
@@ -1,0 +1,5 @@
+package presentation
+
+type Printer interface {
+	Print(interface{}, interface{}) error
+}
