+++ internal/core/presentation/receiver_types.go
@@ -1,4 +1,4 @@
-package usecases
+package presentation
 
 type StdoutPrinter struct {
 	Variant string
@@ -13,9 +13,5 @@ type OpenAIPrinter struct {
 }
 
 type ChartPrinter struct {
-	Variant string
-}
-
-type PineconePrinter struct {
 	Variant string
 }
