+++ internal/core/presentation/writers/chart.go
@@ -1,4 +1,4 @@
-package usecases
+package writers
 
 import (
 	"fmt"
