+++ internal/core/presentation/writers/stdout.go
@@ -1,4 +1,4 @@
-package usecases
+package writers
 
 import (
 	"fmt"
