--- internal/core/usecases/ports.go
@@ -1,5 +1,0 @@
-package usecases
-
-type Printer interface {
-	Write(interface{}, interface{}) error
-}
